+++ src/main/java/mcjty/deepresonance/modules/radiation/manager/RadiationSource.java
@@ -52,7 +52,7 @@ class RadiationSource implements IRadiationSource, INBTSerializable<CompoundNBT>
                     for (int z = (int) (centerZ - radius); z < centerZ + radius; z++) {
                         pos.set(x, y, z);
                         BlockState block = world.getBlockState(pos);
-                        float blocker = RadiationShieldRegistry.getBlocker(block);
+                        float blocker = (float)RadiationShieldRegistry.getBlocker(block);
                         if (blocker < 0.99f) {
                             radiationTree.addBlocker(pos, blocker);
                         }
